requirements.txt
Django
psycopg2-binary>=2.9
djangorestframework
django-cors-headers
djangorestframework-simplejwt